some light refactoring
