--- frontend/src/components/MeetingRoom/VideoSettingsMenu/index.tsx
@@ -1,3 +1,0 @@
-import VideoSettingsMenu from './VideoSettingsMenu';
-
-export default VideoSettingsMenu;
